更改数据类型

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/shop/order/OrderPageListQueryParam.java
@@ -19,7 +19,7 @@ public class OrderPageListQueryParam {
 	public String orderSn;
 	public Byte[] orderStatus;
 	/** 订单类型 */
-	public Integer goodsType;
+	public Byte goodsType;
 	/** 收货人姓名 */
 	public String consignee;
 	public String mobile;
